Refactor: separate keys from lib.rs
